+++ src/runtime/methods.ts
@@ -1,44 +1,20 @@
 import Cookies from 'js-cookie'
 import slugify from '@sindresorhus/slugify'
-import { Ref } from 'vue'
 
-import { useCookieControl } from './composables'
 import { LOCALE_DEFAULT } from './constants'
-import { Cookie, ModuleOptions, Translatable } from './types'
+import { Cookie, Translatable } from './types'
 
-export const useAcceptNecessary = () => {
-  const { cookiesEnabled, moduleOptions } = useCookieControl()
+export const getCookie = (name: string) => Cookies.get(name)
 
-  return () => acceptNecessary(cookiesEnabled, moduleOptions.cookies?.necessary)
-}
+export const getCookieId = (cookie: Cookie) =>
+  cookie.id || slugify(resolveTranslatable(cookie.name))
 
-export const acceptNecessary = (
-  cookiesEnabledRef: Ref<Cookie[]>,
-  cookiesNecessary: Cookie[] = []
-) => {
-  const expires = new Date()
-  expires.setFullYear(expires.getFullYear() + 1)
+export const getCookieIds = (cookies: Cookie[]) =>
+  cookies.map((cookie) => getCookieId(cookie))
 
-  const necessaryCookieIds = cookiesNecessary.map((necessaryCookie) =>
-    getCookieId(necessaryCookie)
-  )
+export const removeCookie = (name: string) => Cookies.remove(name)
 
-  setCookies({
-    isConsentGiven: true,
-    cookieIds: necessaryCookieIds,
-    expires,
-  })
-
-  setHead(cookiesEnabledRef.value)
-  window.location.reload()
-}
-
-export const useResolveTranslatable = (locale = LOCALE_DEFAULT) => {
-  return (translatable: Translatable) =>
-    resolveTranslatable(translatable, locale)
-}
-
-const resolveTranslatable = (
+export const resolveTranslatable = (
   translatable: Translatable,
   locale = LOCALE_DEFAULT
 ) => {
@@ -55,126 +31,13 @@ const resolveTranslatable = (
   return result
 }
 
-export const useSetConsent = () => {
-  const { isConsentGiven, moduleOptions, cookiesEnabled, cookiesEnabledIds } =
-    useCookieControl()
+export const setCookie = (
+  name: string,
+  value: string,
+  options: Cookies.CookieAttributes
+) => Cookies.set(name, value, { sameSite: 'Strict', ...options })
 
-  return () =>
-    setConsent({
-      isInit: false,
-      isConsentGiven,
-      moduleOptions,
-      cookiesEnabled,
-      cookiesEnabledIds,
-    })
-}
-
-export const setConsent = ({
-  isInit = false,
-  isConsentGiven,
-  moduleOptions,
-  cookiesEnabled,
-  cookiesEnabledIds,
-}: {
-  isInit: boolean
-  isConsentGiven: Ref<boolean | undefined>
-  moduleOptions: ModuleOptions
-  cookiesEnabled: Ref<Cookie[]>
-  cookiesEnabledIds: Ref<string[]>
-}) => {
-  isConsentGiven.value = Cookies.get('cookie_control_consent') === 'true'
-  cookiesEnabled.value = []
-  cookiesEnabledIds.value = []
-
-  if (isConsentGiven.value) {
-    const enabledFromCookie = Cookies.get('cookie_control_enabled_cookies')
-
-    cookiesEnabled.value.push(
-      ...moduleOptions.cookies.optional.filter((cookieOptional) =>
-        enabledFromCookie?.includes(getCookieId(cookieOptional))
-      )
-    )
-  }
-
-  if (moduleOptions.cookies?.necessary)
-    cookiesEnabled.value.push(
-      ...moduleOptions.cookies.necessary.filter(
-        (cookieNecessary) => cookieNecessary.src
-      )
-    )
-
-  cookiesEnabledIds.value = cookiesEnabled.value.map((cookieEnabled) =>
-    getCookieId(cookieEnabled)
-  )
-
-  if (process.client && !isInit) {
-    setHead(cookiesEnabled.value)
-    clearCookies(cookiesEnabledIds.value, moduleOptions.cookies.optional)
-  }
-}
-
-export const getCookieControlConsent = () =>
-  Cookies.get('cookie_control_consent')
-
-export const getCookieId = (cookie: Cookie) =>
-  cookie.id || slugify(resolveTranslatable(cookie.name))
-
-export const clearCookies = (
-  cookiesEnabledIds: string[],
-  cookiesOptional: Cookie[]
-) => {
-  const cookiesDisabled = cookiesOptional.filter(
-    (cookieOptional) => !cookiesEnabledIds.includes(getCookieId(cookieOptional))
-  )
-
-  for (const cookieDisabled of cookiesDisabled) {
-    if (!cookieDisabled.targetCookieIds) continue
-
-    for (const cookieDisabledId of cookieDisabled.targetCookieIds) {
-      Cookies.remove(cookieDisabledId)
-    }
-
-    if (cookieDisabled.src) {
-      for (const s of [
-        ...document.head.querySelectorAll(
-          `script[src="${cookieDisabled.src}"]`
-        ),
-      ]) {
-        s.parentNode?.removeChild(s)
-      }
-    }
-  }
-}
-
-export const setHead = (enabledCookies: Cookie[]) => {
-  const head = document.getElementsByTagName('head')[0]
-
-  for (const cookieEnabled of enabledCookies) {
-    if (!cookieEnabled.src) continue
-
-    const script = document.createElement('script')
-    script.src = cookieEnabled.src
-    head.appendChild(script)
-  }
-}
-
-export const setCookies = ({
-  isConsentGiven,
-  cookieIds,
-  expires,
-}: {
-  isConsentGiven: boolean
-  cookieIds: string[]
-  expires: Date
-}) => {
-  Cookies.set(
-    'cookie_control_enabled_cookies',
-    isConsentGiven ? cookieIds.join(',') : '',
-    {
-      expires,
-    }
-  )
-  Cookies.set('cookie_control_consent', isConsentGiven.toString(), {
-    expires,
-  })
+export const useResolveTranslatable = (locale = LOCALE_DEFAULT) => {
+  return (translatable: Translatable) =>
+    resolveTranslatable(translatable, locale)
 }