+++ src/runtime/plugin.ts
@@ -1,14 +1,38 @@
+import Cookies from 'js-cookie'
 import { ref } from 'vue'
-import { setConsent } from './methods'
+
+import { getCookieId } from './methods'
 import { Cookie, State } from './types'
 
 import { defineNuxtPlugin } from '#imports'
 import moduleOptions from '#build/cookie-control-options'
 
 export default defineNuxtPlugin((_nuxtApp) => {
-  const isConsentGiven = ref<boolean>()
-  const cookiesEnabled = ref<Cookie[]>([])
-  const cookiesEnabledIds = ref<string[]>([])
+  const cookieIsConsentGiven = Cookies.get(
+    moduleOptions.cookieNameIsConsentGiven
+  )
+  const cookieCookiesEnabledIds = Cookies.get(
+    moduleOptions.cookieNameCookiesEnabledIds
+  )?.split(',')
+
+  const isConsentGiven = ref<boolean | undefined>(
+    cookieIsConsentGiven === undefined
+      ? undefined
+      : cookieIsConsentGiven === 'true'
+  )
+  const cookiesEnabled = ref<Cookie[] | undefined>(
+    cookieCookiesEnabledIds === undefined
+      ? undefined
+      : [
+          ...moduleOptions.cookies.necessary.filter((cookieNecessary) =>
+            cookieCookiesEnabledIds.includes(getCookieId(cookieNecessary))
+          ),
+          ...moduleOptions.cookies.optional.filter((cookieOptional) =>
+            cookieCookiesEnabledIds.includes(getCookieId(cookieOptional))
+          ),
+        ]
+  )
+  const cookiesEnabledIds = ref<string[] | undefined>(cookieCookiesEnabledIds)
   const isModalActive = ref<boolean>()
 
   const state = {
@@ -18,14 +42,6 @@ export default defineNuxtPlugin((_nuxtApp) => {
     isModalActive,
     moduleOptions,
   } as State
-
-  setConsent({
-    isInit: !process.client,
-    isConsentGiven,
-    moduleOptions,
-    cookiesEnabled,
-    cookiesEnabledIds,
-  })
 
   return {
     provide: {