+++ src/runtime/types.ts
@@ -66,8 +66,12 @@ export interface ModuleOptions {
     necessary: Cookie[]
     optional: Cookie[]
   }
+  cookieExpiryOffsetMs: number
+  cookieNameIsConsentGiven: string
+  cookieNameCookiesEnabledIds: string
   isAcceptNecessaryButtonEnabled?: boolean
   isControlButtonEnabled?: boolean
+  isCookieIdVisible?: boolean
   isCssEnabled?: boolean
   isCssPolyfillEnabled?: boolean
   isDashInDescriptionEnabled?: boolean
@@ -110,8 +114,12 @@ export const DEFAULTS: Required<ModuleOptions> = {
     necessary: [],
     optional: [],
   },
+  cookieExpiryOffsetMs: 1000 * 60 * 60 * 24 * 365, // one year
+  cookieNameIsConsentGiven: 'cookie_control_is_consent_given',
+  cookieNameCookiesEnabledIds: 'cookie_control_cookies_enabled_ids',
   isAcceptNecessaryButtonEnabled: true,
   isControlButtonEnabled: true,
+  isCookieIdVisible: false,
   isCssEnabled: true,
   isCssPolyfillEnabled: true,
   isDashInDescriptionEnabled: true,
@@ -122,9 +130,9 @@ export const DEFAULTS: Required<ModuleOptions> = {
 }
 
 export interface State {
-  cookiesEnabled: Ref<Cookie[]>
-  cookiesEnabledIds: Ref<string[]>
-  isConsentGiven: Ref<boolean>
+  cookiesEnabled: Ref<Cookie[] | undefined>
+  cookiesEnabledIds: Ref<string[] | undefined>
+  isConsentGiven: Ref<boolean | undefined>
   isModalActive: Ref<boolean>
   moduleOptions: ModuleOptions
 }